Splits parser into a module and promotes it to a top-level pass
parser.rs becomes crates/core/src/parser/: a syntax_site core (the SyntaxSite
trait + the single tree-walk engine) plus one file per extractor (labels,
document_date, rheo_var, imports), with mod.rs holding the aggregate
extract_nodes and re-exports.

Moves it out from under reticulate to the crate top level: syntax extraction is
its own pass, distinct from reticulate's spine collection. Updates the handful of
crate::reticulate::parser paths accordingly.
