+++ crates/core/src/reticulate/spine.rs
@@ -2,8 +2,8 @@ use crate::path_utils::{sanitize_handle_segment, to_forward_slash};
 use crate::pdf_utils::DocumentTitle;
 use crate::plugins::SpineOptions;
 use crate::reticulate::bundle_source::BundleSource;
-use crate::reticulate::parser;
-use crate::reticulate::parser::{DocumentDate, RheoValue};
+use crate::parser;
+use crate::parser::{DocumentDate, RheoValue};
 use crate::{Result, RheoError, TYP_EXT};
 use std::collections::{HashMap, HashSet};
 use std::fs;
